book/pyproject.toml
[project]
name = "book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter-book>=2.0.2",
    "evals",
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "numpy>=2.3.5",
    "fonttools>=4.60.1",
]
